// frontend-angular/src/app/api/models/epic.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface Epic {
  createdAt: string;
  description: string | null;
  id: number;
  modifiedAt: string;
  teamId: number;

  /**
   * Drives the disabled-delete affordance (wireframe 5)
   */
  ticketCount: number;
  title: string;
}